Move TestAmple to test module (#4557)
This moves the TestAmple implementation out of the server/base module
test jar and into the test module. TestAmple and associated classes were
originally put there to match the same package as Ample but this does
not work due to jars being sealed. This changes the visibility of some
of the constructors and methods in order to be able to override now that
TestAmple is in a new package.

This is a follow on to #4415
