+++ pom.xml
@@ -624,13 +624,6 @@
         <artifactId>snakeyaml</artifactId>
         <version>2.2</version>
       </dependency>
-      <dependency>
-        <groupId>org.apache.accumulo</groupId>
-        <artifactId>accumulo-server-base</artifactId>
-        <version>${project.version}</version>
-        <type>test-jar</type>
-        <scope>test</scope>
-      </dependency>
     </dependencies>
   </dependencyManagement>
   <dependencies>
